Add CliffEstimator for image processing and pose estimation
- Implemented CliffEstimator class with methods for initialization and estimation.
- Added image preprocessing and bounding box building functions.
- Integrated ONNX Runtime for model inference with CUDA support.
- Updated ModNetMatte and other utility classes to utilize the new OnnxRuntimeCudaProvider for CUDA execution.
- Ensured proper handling of model input and output names.
- Refactored tensor copying logic for output results.

+++ src/utils/SmplifyLite.cpp
@@ -520,8 +520,9 @@ bool SmplifyLiteMultiView(SMPLLayer& smpl_layer,
 
         auto idx_tensor = torch::from_blob(smpl_indices.data(),
                                            {static_cast<int64_t>(smpl_indices.size())},
-                                           torch::kLong)
-                              .clone();
+                                           torch::kInt)
+                              .clone()
+                              .to(torch::kLong);
 
         ViewData vd;
         vd.target_xy = target_xy.to(device);
